Create entrypoint for subnetcalc module
